+++ CHANGELOG.md
@@ -15,6 +15,7 @@ version number is tracked in the file `VERSION`.
 - Extended the lifetime of a `CassResult` into a `Row`. This is a breaking
   change, and may require reworking the code to satisfy the lifetime
   requirement that the `CassResult` must live longer than the `Row`.
+- Implemented `size_hint` on `ResultIterator`.
 
 ### Fixed
  - `CassResult::set_paging_state_token` was implemented incorrectly, namely, it did nothing,